refactor: remove `/utilities/index.ts` file
