+++ src/modules/clients/sagas/togglClientsSagas.ts
@@ -10,7 +10,7 @@ import {
 } from "~/entityOperations/fetchActions";
 import { fetchEntitiesForTool } from "~/entityOperations/fetchEntitiesForTool";
 import { ClientModel, EntityGroup, ToolName } from "~/typeDefs";
-import { validStringify } from "~/utilities";
+import { validStringify } from "~/utilities/textTransforms";
 
 interface TogglClientResponseModel {
   id: number;
