--- src/utilities/index.ts
@@ -1,8 +1,0 @@
-export { getIfDev } from "./getIfDev";
-export {
-  booleanToYesNo,
-  capitalize,
-  getEntityGroupDisplay,
-  kebabCase,
-  validStringify,
-} from "./textTransforms";
